requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
psycopg   # Added as lightweight alternative if needed

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Web Scraping
selenium>=4.21.0
beautifulsoup4==4.12.2
lxml==4.9.3
polars   # Added for data processing

# Google AI Integration
google-generativeai
google-api-python-client

# Email
aiosmtplib==2.0.0

# Environment & Configuration
python-dotenv==1.0.0
loadenv==0.4.0   # Optional: for explicit loadenv usage

# Logging & Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security
bandit==1.7.6
safety==2.3.5

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Development
pre-commit==3.5.0
